根据机型搜索zone&根据name搜索vsw&库存不足时候自动重试

+++ builder/ecs/step_config_vswitch.go
@@ -6,19 +6,20 @@ package ecs
 import (
 	"context"
 	"fmt"
-
-	"github.com/aliyun/alibaba-cloud-sdk-go/services/vpc"
+	"slices"
+	"strings"
 
 	"github.com/aliyun/alibaba-cloud-sdk-go/sdk/responses"
 	"github.com/aliyun/alibaba-cloud-sdk-go/services/ecs"
+	"github.com/aliyun/alibaba-cloud-sdk-go/services/vpc"
 	"github.com/hashicorp/packer-plugin-sdk/multistep"
 	packersdk "github.com/hashicorp/packer-plugin-sdk/packer"
 	"github.com/hashicorp/packer-plugin-sdk/uuid"
 )
 
 type stepConfigAlicloudVSwitch struct {
-	VSwitchId   string
-	ZoneId      string
+	VSwitchIds  []string
+	ZoneIds     []string
 	isCreate    bool
 	CidrBlock   string
 	VSwitchName string
@@ -43,63 +44,17 @@ func (s *stepConfigAlicloudVSwitch) Run(ctx context.Context, state multistep.Sta
 	vpcId := state.Get("vpcid").(string)
 	config := state.Get("config").(*Config)
 
-	if s.ZoneId == "" {
-		describeZonesRequest := ecs.CreateDescribeZonesRequest()
-		describeZonesRequest.RegionId = config.AlicloudRegion
-
-		zonesResponse, err := client.DescribeZones(describeZonesRequest)
-		if err != nil {
-			return halt(state, err, "Query for available zones failed")
+	// 指定交换机
+	if s.VSwitchIds[0] != "" {
+		if len(s.ZoneIds) != 1 {
+			halt(state, fmt.Errorf("指定vswitch_id必须指定唯一zone_id"), "")
+			return multistep.ActionHalt
 		}
-
-		var instanceTypes []string
-		zones := zonesResponse.Zones.Zone
-		for _, zone := range zones {
-			isVSwitchSupported := false
-			for _, resourceType := range zone.AvailableResourceCreation.ResourceTypes {
-				if resourceType == "VSwitch" {
-					isVSwitchSupported = true
-				}
-			}
-
-			if isVSwitchSupported {
-				for _, instanceType := range zone.AvailableInstanceTypes.InstanceTypes {
-					if instanceType == config.InstanceType {
-						s.ZoneId = zone.ZoneId
-						break
-					}
-					instanceTypes = append(instanceTypes, instanceType)
-				}
-			}
-		}
-
-		if s.ZoneId == "" {
-			if len(instanceTypes) > 0 {
-				ui.Say(fmt.Sprintf("The instance type %s isn't available in this region."+
-					"\n You can either change the instance to one of following: %v \n"+
-					"or choose another region.", config.InstanceType, instanceTypes))
-
-				state.Put("error", fmt.Errorf("The instance type %s isn't available in this region."+
-					"\n You can either change the instance to one of following: %v \n"+
-					"or choose another region.", config.InstanceType, instanceTypes))
-				return multistep.ActionHalt
-			} else {
-				ui.Say(fmt.Sprintf("The instance type %s isn't available in this region."+
-					"\n You can change to other regions.", config.InstanceType))
-
-				state.Put("error", fmt.Errorf("The instance type %s isn't available in this region."+
-					"\n You can change to other regions.", config.InstanceType))
-				return multistep.ActionHalt
-			}
-		}
-	}
-
-	if len(s.VSwitchId) != 0 || len(s.VSwitchName) != 0 {
 		describeVSwitchesRequest := vpc.CreateDescribeVSwitchesRequest()
 		describeVSwitchesRequest.VpcId = vpcId
-		describeVSwitchesRequest.VSwitchId = s.VSwitchId
+		describeVSwitchesRequest.VSwitchId = s.VSwitchIds[0]
 		describeVSwitchesRequest.VSwitchName = s.VSwitchName
-		describeVSwitchesRequest.ZoneId = s.ZoneId
+		describeVSwitchesRequest.ZoneId = s.ZoneIds[0]
 
 		vswitchesResponse, err := vpcClient.DescribeVSwitches(describeVSwitchesRequest)
 		if err != nil {
@@ -108,13 +63,75 @@ func (s *stepConfigAlicloudVSwitch) Run(ctx context.Context, state multistep.Sta
 
 		vswitch := vswitchesResponse.VSwitches.VSwitch
 		if len(vswitch) > 0 {
-			state.Put("vswitchid", vswitch[0].VSwitchId)
+			state.Put("vswitches", vswitch)
 			s.isCreate = false
 			return multistep.ActionContinue
 		}
 
 		s.isCreate = false
-		return halt(state, fmt.Errorf("The specified vswitch {%s} doesn't exist.", s.VSwitchId), "")
+		return halt(state, fmt.Errorf("the specified vswitch {%s} doesn't exist", s.VSwitchIds[0]), "")
+	}
+
+	// 根据机型自动选择可用区
+	if s.ZoneIds[0] == "" {
+		ui.Say("Searching zones...")
+		availableResourceRequest := ecs.CreateDescribeAvailableResourceRequest()
+		availableResourceRequest.RegionId = config.AlicloudRegion
+		availableResourceRequest.DestinationResource = "InstanceType"
+		availableResourceRequest.IoOptimized = "optimized"
+		availableResourceRequest.ResourceType = "instance"
+		availableResourceRequest.InstanceType = config.InstanceType
+
+		resourceResponse, err := client.DescribeAvailableResource(availableResourceRequest)
+		if err != nil {
+			return halt(state, err, "Query for available instance zones failed")
+		}
+
+		zones := make([]string, 0)
+		for _, zone := range resourceResponse.AvailableZones.AvailableZone {
+			if zone.Status == "Available" &&
+				zone.AvailableResources.AvailableResource[0].SupportedResources.SupportedResource[0].Status == "Available" {
+				zones = append(zones, zone.ZoneId)
+			}
+		}
+		if len(zones) == 0 {
+			ui.Say(fmt.Sprintf("实例类型 %s 没有在售可用区", config.InstanceType))
+			state.Put("error", fmt.Errorf("实例类型 %s 没有在售可用区", config.InstanceType))
+			return multistep.ActionHalt
+		}
+		s.ZoneIds = zones
+		ui.Say("Candidate zones are: " + strings.Join(s.ZoneIds, ", "))
+	}
+
+	// 根据机型自动选择可用区和交换机
+	if len(s.VSwitchName) != 0 {
+		ui.Say(fmt.Sprintf("Searching vswitches using name: %s ...", s.VSwitchName))
+		s.isCreate = false
+		// 搜索机型在售所有可用区内符合subnet名称的subnet
+		// 由于不可以指定可用区列表，因此需要遍历返回值然后进行过滤
+		describeVSwitchesRequest := vpc.CreateDescribeVSwitchesRequest()
+		describeVSwitchesRequest.VpcId = vpcId
+		describeVSwitchesRequest.VSwitchName = s.VSwitchName
+
+		vSwitchesResponse, err := vpcClient.DescribeVSwitches(describeVSwitchesRequest)
+		if err != nil {
+			return halt(state, err, "Failed querying vswitch")
+		}
+
+		vSwitchCandidates := vSwitchesResponse.VSwitches.VSwitch
+		if len(vSwitchCandidates) > 0 {
+			vSwitches := make([]vpc.VSwitch, 0)
+			for _, v := range vSwitchCandidates {
+				if slices.Contains(s.ZoneIds, v.ZoneId) {
+					vSwitches = append(vSwitches, v)
+				}
+			}
+			state.Put("vswitches", vSwitches)
+
+			return multistep.ActionContinue
+		}
+
+		return halt(state, fmt.Errorf("The specified vswitch {%s} doesn't exist.", s.VSwitchName), "")
 	}
 
 	if config.CidrBlock == "" {
@@ -122,11 +139,22 @@ func (s *stepConfigAlicloudVSwitch) Run(ctx context.Context, state multistep.Sta
 	}
 
 	ui.Say("Creating vswitch...")
+	// 由于cidr段无法有效指定多个，因此建立vsw只建立第一个可用区
+	s.isCreate = true
+	vSwitchChildren := make([]vpc.VSwitch, 0)
+	vSwitchIds := make([]string, 0)
+	zoneId := s.ZoneIds[0]
 
-	createVSwitchRequest := s.buildCreateVSwitchRequest(state)
+	ui.Say("Try to create vsw in zone: " + zoneId)
+	createVSwitchRequest := vpc.CreateCreateVSwitchRequest()
+	createVSwitchRequest.ClientToken = uuid.TimeOrderedUUID()
+	createVSwitchRequest.CidrBlock = s.CidrBlock
+	createVSwitchRequest.ZoneId = zoneId
+	createVSwitchRequest.VpcId = vpcId
+	createVSwitchRequest.VSwitchName = s.VSwitchName
 	createVSwitchResponse, err := client.WaitForExpected(&WaitForExpectArgs{
 		RequestFunc: func() (responses.AcsResponse, error) {
-			return client.CreateVSwitch(createVSwitchRequest)
+			return vpcClient.CreateVSwitch(createVSwitchRequest)
 		},
 		EvalFunc: client.EvalCouldRetryResponse(createVSwitchRetryErrors, EvalRetryErrorType),
 	})
@@ -134,31 +162,33 @@ func (s *stepConfigAlicloudVSwitch) Run(ctx context.Context, state multistep.Sta
 		return halt(state, err, "Error Creating vswitch")
 	}
 
-	vSwitchId := createVSwitchResponse.(*ecs.CreateVSwitchResponse).VSwitchId
+	vSwitchId := createVSwitchResponse.(*vpc.CreateVSwitchResponse).VSwitchId
 
-	describeVSwitchesRequest := ecs.CreateDescribeVSwitchesRequest()
+	describeVSwitchesRequest := vpc.CreateDescribeVSwitchesRequest()
 	describeVSwitchesRequest.VpcId = vpcId
 	describeVSwitchesRequest.VSwitchId = vSwitchId
 
 	_, err = client.WaitForExpected(&WaitForExpectArgs{
 		RequestFunc: func() (responses.AcsResponse, error) {
-			return client.DescribeVSwitches(describeVSwitchesRequest)
+			return vpcClient.DescribeVSwitches(describeVSwitchesRequest)
 		},
 		EvalFunc: func(response responses.AcsResponse, err error) WaitForExpectEvalResult {
 			if err != nil {
 				return WaitForExpectToRetry
 			}
 
-			vSwitchesResponse := response.(*ecs.DescribeVSwitchesResponse)
+			vSwitchesResponse := response.(*vpc.DescribeVSwitchesResponse)
 			vSwitches := vSwitchesResponse.VSwitches.VSwitch
 			if len(vSwitches) > 0 {
 				for _, vSwitch := range vSwitches {
 					if vSwitch.Status == VSwitchStatusAvailable {
+						vSwitchChildren = append(vSwitchChildren, vSwitch)
+						vSwitchIds = append(vSwitchIds, vSwitch.VSwitchId)
+						s.VSwitchIds = vSwitchIds // 保证销毁时可以获取到vswitch id
 						return WaitForExpectSuccess
 					}
 				}
 			}
-
 			return WaitForExpectToRetry
 		},
 		RetryTimes: shortRetryTimes,
@@ -167,11 +197,9 @@ func (s *stepConfigAlicloudVSwitch) Run(ctx context.Context, state multistep.Sta
 	if err != nil {
 		return halt(state, err, "Timeout waiting for vswitch to become available")
 	}
-
 	ui.Message(fmt.Sprintf("Created vswitch: %s", vSwitchId))
-	state.Put("vswitchid", vSwitchId)
-	s.isCreate = true
-	s.VSwitchId = vSwitchId
+
+	state.Put("vswitches", vSwitchChildren)
 	return multistep.ActionContinue
 }
 
@@ -185,30 +213,22 @@ func (s *stepConfigAlicloudVSwitch) Cleanup(state multistep.StateBag) {
 	client := state.Get("client").(*ClientWrapper)
 	ui := state.Get("ui").(packersdk.Ui)
 
-	_, err := client.WaitForExpected(&WaitForExpectArgs{
-		RequestFunc: func() (responses.AcsResponse, error) {
-			request := ecs.CreateDeleteVSwitchRequest()
-			request.VSwitchId = s.VSwitchId
-			return client.DeleteVSwitch(request)
-		},
-		EvalFunc:   client.EvalCouldRetryResponse(deleteVSwitchRetryErrors, EvalRetryErrorType),
-		RetryTimes: shortRetryTimes,
-	})
+	for _, id := range s.VSwitchIds {
+		if len(id) == 0 {
+			continue
+		}
+		_, err := client.WaitForExpected(&WaitForExpectArgs{
+			RequestFunc: func() (responses.AcsResponse, error) {
+				request := ecs.CreateDeleteVSwitchRequest()
+				request.VSwitchId = id
+				return client.DeleteVSwitch(request)
+			},
+			EvalFunc:   client.EvalCouldRetryResponse(deleteVSwitchRetryErrors, EvalRetryErrorType),
+			RetryTimes: shortRetryTimes,
+		})
 
-	if err != nil {
-		ui.Error(fmt.Sprintf("Error deleting vswitch, it may still be around: %s", err))
+		if err != nil {
+			ui.Error(fmt.Sprintf("Error deleting vswitch, it may still be around: %s", err))
+		}
 	}
-}
-
-func (s *stepConfigAlicloudVSwitch) buildCreateVSwitchRequest(state multistep.StateBag) *ecs.CreateVSwitchRequest {
-	vpcId := state.Get("vpcid").(string)
-
-	request := ecs.CreateCreateVSwitchRequest()
-	request.ClientToken = uuid.TimeOrderedUUID()
-	request.CidrBlock = s.CidrBlock
-	request.ZoneId = s.ZoneId
-	request.VpcId = vpcId
-	request.VSwitchName = s.VSwitchName
-
-	return request
 }
